Improve the mod *Fixed A LOT of crashes *Changed the uuid and the player name methods *Fixed code structure, removed unused code *Added README.md and updated the LICENSE *Updated the fabric loom version + gradle version *Fixed version name
