+++ gradle.properties
@@ -9,9 +9,9 @@ yarn_mappings=1.20.4+build.3
 loader_version=0.15.3
 
 # Mod Properties
-mod_version=1.2.0
-maven_group=net.uhb217.playertrackermod
-archives_base_name=playertrackermod
+mod_version=0.2
+maven_group=net.uhb217.playertracker
+archives_base_name=Player-Tracker-fabric-1.20.4
 
 # Dependencies
 fabric_version=0.92.0+1.20.4
