+++ src/client/java/net/uhb217/playertracker/utils/NBTConfigUtils.java
@@ -1,0 +1,7 @@
+package net.uhb217.playertracker.utils;
+
+import net.minecraft.nbt.NbtCompound;
+
+public interface NBTConfigUtils {
+    NbtCompound playerTracker$getPersistentData();
+}
