--- src/client/java/net/uhb217/playertrackermod/mixin/client/ModEntityDataSaverMixin.java
@@ -1,37 +1,0 @@
-package net.uhb217.playertrackermod.mixin.client;
-
-import net.minecraft.entity.Entity;
-import net.minecraft.nbt.NbtCompound;
-import net.uhb217.playertrackermod.utils.IEntityDataSaver;
-import org.spongepowered.asm.mixin.Mixin;
-import org.spongepowered.asm.mixin.injection.At;
-import org.spongepowered.asm.mixin.injection.Inject;
-import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
-import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
-
-@Mixin(Entity.class)
-public abstract class ModEntityDataSaverMixin implements IEntityDataSaver {
-    private NbtCompound persistentData;
-
-    @Override
-    public NbtCompound getPersistentData() {
-        if(this.persistentData == null) {
-            this.persistentData = new NbtCompound();
-        }
-        return persistentData;
-    }
-
-    @Inject(method = "writeNbt", at = @At("HEAD"))
-    protected void injectWriteMethod(NbtCompound nbt, CallbackInfoReturnable info) {
-        if(persistentData != null) {
-            nbt.put("glowingentity.uhb_data", persistentData);
-        }
-    }
-
-    @Inject(method = "readNbt", at = @At("HEAD"))
-    protected void injectReadMethod(NbtCompound nbt, CallbackInfo info) {
-        if (nbt.contains("glowingentity.uhb_data", 10)) {
-            persistentData = nbt.getCompound("glowingentity.uhb_data");
-        }
-    }
-}
