--- src/client/java/net/uhb217/playertrackermod/utils/IEntityDataSaver.java
@@ -1,7 +1,0 @@
-package net.uhb217.playertrackermod.utils;
-
-import net.minecraft.nbt.NbtCompound;
-
-public interface IEntityDataSaver {
-    NbtCompound getPersistentData();
-}
