+++ src/client/resources/playertracker.client.mixins.json
@@ -1,0 +1,15 @@
+{
+  "required": true,
+  "package": "net.uhb217.playertracker.mixin.client",
+  "compatibilityLevel": "JAVA_17",
+  "client": [
+    "ModelPredicateProviderMixin",
+    "ModEntityDataSaverMixin"
+  ],
+  "injectors": {
+    "defaultRequire": 1
+  },
+  "mixins": [
+
+  ]
+}
