--- src/client/resources/playertrackermod.client.mixins.json
@@ -1,16 +1,0 @@
-{
-  "required": true,
-  "package": "net.uhb217.playertrackermod.mixin.client",
-  "compatibilityLevel": "JAVA_17",
-  "client": [
-    "EntityRendererMixin",
-    "ModelPredicateProviderMixin",
-    "ModEntityDataSaverMixin"
-  ],
-  "injectors": {
-    "defaultRequire": 1
-  },
-  "mixins": [
-
-  ]
-}
