+++ src/main/resources/playertracker.mixins.json
@@ -1,0 +1,11 @@
+{
+  "required": true,
+  "package": "net.uhb217.playertracker.mixin",
+  "compatibilityLevel": "JAVA_17",
+  "mixins": [
+    "ExampleMixin"
+  ],
+  "injectors": {
+    "defaultRequire": 1
+  }
+}
