--- src/main/resources/playertrackermod.mixins.json
@@ -1,11 +1,0 @@
-{
-  "required": true,
-  "package": "net.uhb217.playertrackermod.mixin",
-  "compatibilityLevel": "JAVA_17",
-  "mixins": [
-    "ExampleMixin"
-  ],
-  "injectors": {
-    "defaultRequire": 1
-  }
-}
